Hackathon/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==2.0.0
openai==1.35.10
pandas==2.2.2
plum-dispatch==1.7.4
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
PySimpleGUI==5.0.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.5.15
requests==2.32.3
rsa==4.9
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
spire==0.4.2
Spire.Doc==12.6.0
tiktoken==0.7.0
tk==0.1.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2